+++ src/c/main.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "clay_settings.h"
+#include "settings/clay_settings.h"
 
 // Rebuild the layer layout from the current settings and reload all layers.
 // Call this when row layout settings change at runtime.
-void main_reload_layout(ClaySettings *settings);
+void main_reload_layout(ClaySettings * settings);
