+++ src/c/system/battery.h
