+++ src/c/system/battery_listener.h
