+++ src/c/ui/layer/date.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "../layer_factory.h"
+#include "../ui/layer_factory.h"
 
 // update all date layer instances (backward compatible wrapper, called by listeners)
 void update_date();
