+++ src/c/ui/layer/dotted_text_layer.c
@@ -1,6 +1,6 @@
-#include "dotted_text_layer.h"
-#include "pixel_matrix_drawer.h"
-#include "clay_settings.h"
+#include "../dotted_text_layer.h"
+#include "../graphics/pixel_matrix_drawer.h"
+#include "../settings/clay_settings.h"
 
 static DottedTextLayerData *get_layer_data(const DottedTextLayer *dotted_text_layer) {
     return layer_get_data(dotted_text_layer);
