+++ src/c/ui/layer/heartrate.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "../layer_factory.h"
+#include "../ui/layer_factory.h"
 
 // update all heartrate layer instances (called by listeners)
 void update_heartrate();
@@ -8,5 +8,4 @@ void update_heartrate();
 Layer *create_heartrate_layer(LayerBuilder builder);
 
 // destroy the layer
-void destroy_heartrate_layer(Layer *layer);
-
+void destroy_heartrate_layer(Layer * layer);
