+++ src/c/ui/layer/stepcount.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "../layer_factory.h"
+#include "../ui/layer_factory.h"
 
 // update all stepcount layer instances (backward compatible wrapper, called by listeners)
 void update_stepcount();
