+++ src/c/ui/layer/weather.c
@@ -2,10 +2,10 @@
 #include <string.h>
 #include <stdlib.h>
 #include "../developer_options.h"
-#include "../theme.h"
-#include "../clay_settings.h"
-#include "../dotted_text_layer.h"
-#include "../layer_factory.h"
+#include "../ui/theme.h"
+#include "../settings/clay_settings.h"
+#include "dotted_text_layer.h"
+#include "../ui/layer_factory.h"
 #include "weather_forecast.h"
 
 #define MAX_WEATHER_LAYERS 7
