+++ src/c/ui/layer/weather.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <pebble.h>
-#include "../layer_factory.h"
-#include "../persist_keys.h"
+#include "../ui/layer_factory.h"
+#include "../settings/persist_keys.h"
 
 // Persistent storage key
 #define WEATHER_DATA_KEY PERSIST_KEY_WEATHER_DATA
