+++ src/c/ui/layer/weather_forecast.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../layer_factory.h"
+#include "../ui/layer_factory.h"
 
 void update_temperature_forecast();
 Layer *create_temperature_forecast_layer(LayerBuilder builder);
