+++ tests/c/clay_settings_test.c
@@ -8,7 +8,7 @@
 
 // 2. Include the headers and source file to be tested.
 #include "../../src/c/clay_settings.h"
-#include "../../src/c/clay_settings.c"
+#include "../../src/c/settings/clay_settings.c"
 
 // --- Mocks ---
 void app_messaging_request_settings() {
